webapp/pyproject.toml
[tool.poetry]
name = "webapp"
version = "0.1.0"
description = ""
authors = ["L3vi-Ackerman <l3vi3211@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
psycopg2-binary = "^2.9.10"
dotenv = "^0.9.9"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
django-configurations = "^2.5.1"
pillow = "^11.3.0"
channels = {extras = ["daphne"], version = "^4.2.2"}
channels-redis = "^4.2.1"
hashids = "^1.3.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
